Silence another Clang -Wsign-conversion warning

+++ breath/diagnostics/brt/exception.cpp
@@ -8,6 +8,7 @@
 
 #include "breath/diagnostics/exception.hpp"
 #include "breath/counting/count.hpp"
+#include <cstddef>
 #include <cstring>
 
 namespace breath {
@@ -30,7 +31,8 @@ exception::exception( std::string const & what_msg ) noexcept
     std::strcpy( curr, sep ) ;
     curr += ( count( sep ) - 1 ) ;
     std::strncpy( curr, what_msg.c_str(),
-        what_message_max_size - (curr - &m_what_message[ 0 ] ) ) ;
+        static_cast< std::size_t >( what_message_max_size -
+                                    (curr - &m_what_message[ 0 ] ) ) ) ;
 }
 
 char const *
